% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catatab.R
\name{catatab}
\alias{catatab}
\title{CATA Table}
\usage{
catatab(df, x, stub)
}
\arguments{
\item{df}{A data frame}

\item{x}{A variable in the data frame}
}
\value{
A nicely formatted table
}
\description{
Creates a nicely formatted frequency table with percentage and n in parentheses.
}
\examples{
frtab(df, "Q23")

}
