% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbtab.R
\name{cbtab}
\alias{cbtab}
\title{Frequency Table}
\usage{
cbtab(df, x)
}
\arguments{
\item{df}{A data frame}

\item{x}{The column number of a variable in df}
}
\value{

}
\description{
Creates a nicely formatted frequency table with percentage and n in parentheses.  Best used to create a codebook using a for loop.
}
\details{
Here's a whole bunch of other things
}
\examples{
# Table of infert data variable education in column 1
cbtab(infert, 1)
}
