% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebk.R
\name{codebk}
\alias{codebk}
\title{Title}
\usage{
codebk(x, df, varlist)
}
\arguments{
\item{x}{A file name to export report to}

\item{df}{A data file}

\item{varlist}{A list of variable names and labels}
}
\value{

}
\description{
Title
}
\examples{
codebk("P9999 Frequency Report.qmd", df, varlist)
}
