\name{coef.midrq}
\alias{coefficients.midrq}
\alias{coef.midrq}

\title{
Extract Coefficients
}

\description{
\code{coef} extracts model coefficients from \code{midrq} objects.
}

\usage{
\method{coef}{midrq}(object, ...)
\method{coefficients}{midrq}(object, ...)
}

\arguments{
  \item{object}{
an \code{midrq} object.
}
  \item{\dots}{
not used.
}
}

\value{
a vector for single quantiles or a matrix for multiple quantiles. 
}

\author{
Marco Geraci
}

\seealso{
\code{\link{midrq}}
}

\keyword{coefficients}
