\name{coef.qrr}
\alias{coefficients.qrr}
\alias{coef.qrr}

\title{
Extract Coefficients
}

\description{
\code{coef} extracts model coefficients from \code{qrr} objects.
}

\usage{
\method{coef}{qrr}(object, ...)
\method{coefficients}{qrr}(object, ...)
}

\arguments{
  \item{object}{
an object of \code{\link{class}} \code{qrr}.
}
  \item{\dots}{
not used.
}
}

\value{
a vector of estimated coefficients. 
}

\author{
Marco Geraci
}

\seealso{
\code{\link{qrr}}
}

\keyword{coefficients}
