\name{coef.rq.bin}
\alias{coefficients.rq.bin}
\alias{coef.rq.bin}

\title{
Extract Coefficients
}

\description{
\code{coef} extracts model coefficients from \code{rq.bin} objects.
}

\usage{
\method{coef}{rq.bin}(object, ...)
\method{coefficients}{rq.bin}(object, ...)
}

\arguments{
  \item{object}{
an \code{rq.bin} object.
}
  \item{\dots}{
not used.
}
}

\value{
a vector for single quantiles or a matrix for multiple quantiles. 
}

\author{
Marco Geraci
}

\seealso{
\code{\link{rq.bin}}
}

\keyword{coefficients}
