\name{coef.rqt}
\alias{coefficients.rqt}
\alias{coef.rqt}

\title{
Extract Coefficients
}

\description{
\code{coef} extracts model coefficients from \code{rqt} objects.
}

\usage{
\method{coef}{rqt}(object, ...)
\method{coefficients}{rqt}(object, ...)
}

\arguments{
  \item{object}{
an \code{rqt} object.
}
  \item{\dots}{
not used.
}
}

\value{
a vector for single quantiles or a matrix for multiple quantiles. 
}

\author{
Marco Geraci
}

\keyword{coefficients}
