% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctab.R
\name{ctab}
\alias{ctab}
\title{Title}
\usage{
ctab(df, x, y)
}
\arguments{
\item{df}{A data frame}

\item{x}{A row variable}

\item{y}{A col variable}
}
\value{
A nicely formatted table of x by y with output formatted as \% (N)
}
\description{
Title
}
\examples{
ctab(df, Q2, Q5)
}
