\name{esterase}
\alias{esterase}
\title{Esterase Essay Data}
\description{
  The \code{esterase} data frame has 113 rows and 2 columns with the
  results of an essay for the concentration of an enzyme esterase.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Esterase}{
      amount of esterase.
    }
    \item{Count}{
      observed count.
    }

  }
}

\details{
The esterase essay data were reported by Caroll and Ruppert (1988) and successively analyzed by Zhao (2000).
}

\references{
Zhao QS. Restricted regression quantiles. Journal of Multivariate Analysis 2000;72(1):78-99.
}

\source{
R. J. Carroll and D. Ruppert, Transformation and Weighting in Regression. London: Chapman and Hall, 1988.
}


\keyword{datasets}
\keyword{ quantile crossing }
