\name{fitted.midrq}
\alias{fitted.midrq}

\title{
Extract Fitted Values from Mid-Quantile Transformation Models
}

\description{
This function extracts fitted values from objects of class \code{midrq}.
}

\usage{
\method{fitted}{midrq}(object, ...)
}

\arguments{
  \item{object}{
an object of \code{\link{class}} \code{midrq}.
}
  \item{\dots}{
other arguments.
}
}

\value{
a vector or a matrix or an array of fitted values.
}

\author{
Marco Geraci
}

\seealso{
\code{\link{predict.midrq}}
}


\keyword{ predict }

