\name{fitted.rq.bin}
\alias{fitted.rq.bin}

\title{
Extract Fitted Values from Binary Quantile Regression Models
}

\description{
This function extracts fitted values from objects of class \code{rq.bin}.
}

\usage{
\method{fitted}{rq.bin}(object, ...)
}

\arguments{
  \item{object}{
an object of \code{\link{class}} \code{rq.bin}.
}
  \item{\dots}{
other arguments.
}
}

\value{
a vector or a matrix or an array of fitted values.
}

\author{
Marco Geraci
}

\seealso{
\code{\link{predict.rq.bin}}
}


\keyword{ predict }

