\name{fitted.rq.counts}
\alias{fitted.rq.counts}

\title{
Extract Fitted Values from Quantile Regression Models for Counts
}

\description{
This function extracts fitted values from objects of class \code{rq.counts}.
}

\usage{
\method{fitted}{rq.counts}(object, ...)
}

\arguments{
  \item{object}{
an object of \code{\link{class}} \code{rq.counts}.
}
  \item{\dots}{
other arguments.
}
}

\value{
a vector or a matrix or an array of fitted values.
}

\author{
Marco Geraci
}

\seealso{
\code{\link{predict.rq.counts}}
}


\keyword{ predict }

