\name{fitted.rqt}
\alias{fitted.rqt}

\title{
Extract Fitted Values from Quantile Regression Transformation Models
}

\description{
This function extracts fitted values from objects of class \code{rqt}.
}

\usage{
\method{fitted}{rqt}(object, ...)
}

\arguments{
  \item{object}{
an object of \code{\link{class}} \code{rqt}.
}
  \item{\dots}{
other arguments.
}
}

\value{
a vector or a matrix or an array of fitted values.
}

\author{
Marco Geraci
}

\seealso{
\code{\link{predict.rqt}}
}


\keyword{ predict }

