% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbtab.R
\name{frtab1}
\alias{frtab1}
\title{Frequency Table}
\usage{
frtab1(df, x)
}
\arguments{
\item{df}{A data frame}

\item{x}{The column number of a variable in df}
}
\value{

}
\description{
Creates a nicely formatted frequency table with percentage and n in parentheses.  Best used to create a codebook using a for loop.
}
\details{
Here's a whole bunch of other things
}
\examples{
# Table of airquality data variable Month in column 5
frtab1(airquality, 5)
}
