\name{print.cmidecdf}
\alias{print.cmidecdf}

\title{
Print Mid-distribution Functions
}
\description{
Print an object generated by \code{\link{cmidecdf}}.
}
\usage{
\method{print}{cmidecdf}(x, ...)
}

\arguments{
  \item{x}{
a \code{midecdf} or a \code{midquantile} object.
}
  \item{\dots}{
not used.
}
}


\author{
Marco Geraci
}


\seealso{
\code{\link{cmidecdf}}
}

\keyword{print}
