\name{print.dqc}
\alias{print.dqc}

\title{
Print Directional Quantile Classification Objects
}

\description{
Print an object of class \code{dqc}.
}

\usage{
\method{print}{dqc}(x, ...)
}

\arguments{
  \item{x}{
an object of \code{\link{class}} \code{dqc}.
}
  \item{\dots}{
other arguments used by \code{\link{print.default}}.
}
}



\author{
Marco Geraci
}

\seealso{

\code{\link{dqc}}
}

\keyword{ print }

