\name{print.GOFTest}
\alias{print.GOFTest}

\title{
Print Goodness-of-Fit Test for Quantile Regression Models
}
\description{
Print an object generated by \code{\link{GOFTest}}.
}
\usage{
\method{print}{GOFTest}(x, digits = max(3, getOption("digits") - 3), ...)
}

\arguments{
  \item{x}{
an \code{GOFTest} object.
}
  \item{digits}{
a non-null value for digits specifies the minimum number of significant digits to be printed in values.
}

  \item{\dots}{
not used.
}
}


\author{
Marco Geraci
}


\seealso{
\code{\link{GOFTest}}
}

\keyword{print}
