\name{print.midquantile}
\alias{print.midecdf}
\alias{print.midquantile}

\title{
Print Mid-distribution Functions
}
\description{
Print an object generated by \code{\link{midecdf}} or \code{\link{midquantile}}.
}
\usage{
\method{print}{midecdf}(x, ...)
\method{print}{midquantile}(x, ...)
}

\arguments{
  \item{x}{
a \code{midecdf} or a \code{midquantile} object.
}
  \item{\dots}{
not used.
}
}


\author{
Marco Geraci
}


\seealso{
\code{\link{midecdf}}, \code{\link{midquantile}}
}

\keyword{print}
