\name{print.midrq}
\alias{print.summary.midrq}
\alias{print.midrq}

\title{
Print Mid-Quantile Models
}

\description{
Print an object of class \code{midrq} or \code{summary.midrq}.
}

\usage{
\method{print}{midrq}(x, ...)
\method{print}{summary.midrq}(x, ...)
}

\arguments{
  \item{x}{
an object of \code{\link{class}} \code{midrq} or \code{summary.midrq}.
}
  \item{\dots}{
other arguments used by \code{\link{print.default}}.
}
}



\author{
Marco Geraci
}

\seealso{

\code{\link{midrq}}
}

\keyword{ print }

