\name{print.qlss}
\alias{print.qlss}

\title{
Print Quantile-based Summary Statistics for Location, Scale and Shape
}
\description{
Print an object generated by \code{\link{qlss}}.
}
\usage{
\method{print}{qlss}(x, ...)
}

\arguments{
  \item{x}{
an \code{qlss} object.
}
  \item{\dots}{
not used.
}
}


\author{
Marco Geraci
}


\seealso{
\code{\link{qlss}}
}

\keyword{print}
