\name{print.qrr}
\alias{print.summary.qrr}
\alias{print.qrr}

\title{
Print Quantile Ratio Regression Models
}

\description{
Print an object of class \code{qrr} or \code{summary.qrr}.
}

\usage{
\method{print}{qrr}(x, ...)
\method{print}{summary.qrr}(x, ...)
}

\arguments{
  \item{x}{
an object of \code{\link{class}} \code{qrr} or \code{summary.qrr}.
}
  \item{\dots}{
other arguments used by \code{\link{print.default}}.
}
}



\author{
Marco Geraci
}

\seealso{
\code{\link{qrr}}
}


\keyword{ print }

