\name{print.rq.bin}
\alias{print.rq.bin}

\title{
Print Binary Quantile Regression Models
}

\description{
Print an object of class \code{rq.bin}.
}

\usage{
\method{print}{rq.bin}(x, digits = max(6, getOption("digits")), ...)
}

\arguments{
  \item{x}{
an object of \code{\link{class}} \code{rq.bin}.
}
  \item{digits}{
a non-null value for digits specifies the minimum number of significant digits to be printed in values.
}

  \item{\dots}{
not used.
}
}



\author{
Marco Geraci
}

\seealso{

\code{\link{rq.bin}}
}

\keyword{ print }

