\name{print.rq.counts}
\alias{print.rq.counts}

\title{
Print rq.counts
}
\description{
Print an object generated by \code{\link{rq.counts}}.
}
\usage{
\method{print}{rq.counts}(x, digits = max(3, getOption("digits") - 3), ...)
}

\arguments{
  \item{x}{
an \code{rq.counts} object.
}
  \item{digits}{
a non-null value for digits specifies the minimum number of significant digits to be printed in values.
}

  \item{\dots}{
not used.
}
}


\author{
Marco Geraci
}


\seealso{
\code{\link{rq.counts}}
}

\keyword{print}
