\name{print.rqt}
\alias{print.summary.rqt}
\alias{print.rqt}

\title{
Print Transformation Models
}

\description{
Print an object of class \code{rqt} or \code{summary.rqt}.
}

\usage{
\method{print}{rqt}(x, ...)
\method{print}{summary.rqt}(x, ...)
}

\arguments{
  \item{x}{
an object of \code{\link{class}} \code{rqt} or \code{summary.rqt}.
}
  \item{\dots}{
other arguments used by \code{\link{print.default}}.
}
}



\author{
Marco Geraci
}

\seealso{

\code{\link{tsrq}}, \code{\link{rcrq}}, \code{\link{tsrq2}} or \code{\link{nlrq2}}
}

\keyword{ print }

