\name{print.rrq}
\alias{print.summary.rrq}
\alias{print.rrq}

\title{
Print Restricted Quantile Regression Models
}

\description{
Print an object of class \code{rrq} or \code{summary.rrq}.
}

\usage{
\method{print}{rrq}(x, ...)
\method{print}{summary.rrq}(x, ...)
}

\arguments{
  \item{x}{
an object of \code{\link{class}} \code{rrq} or \code{summary.rrq}.
}
  \item{\dots}{
other arguments used by \code{\link{print.default}}.
}
}



\author{
Marco Geraci
}

\seealso{
\code{\link{rrq}}
}


\keyword{ print }

