\name{print.tsrq}
\alias{print.tsrq2}
\alias{print.rcrq}
\alias{print.nlrq2}
\alias{print.tsrq}

\title{
Print Transformation Models
}

\description{
Print an object generated by \code{\link{tsrq}}, \code{\link{tsrq2}} or \code{\link{rcrq}}.
}

\usage{
\method{print}{tsrq}(x, ...)
\method{print}{tsrq2}(x, ...)
\method{print}{rcrq}(x, ...)
\method{print}{nlrq2}(x, ...)
}

\arguments{
  \item{x}{
an object of \code{\link{class}} \code{tsrq}, \code{tsrq2} or \code{nlrq2}.
}
  \item{\dots}{
other arguments used by \code{\link{print.default}}.
}
}



\author{
Marco Geraci
}


\keyword{ print }

