\name{Qtools-internal}
\title{Internal Qtools objects}
\alias{.First.lib}
\alias{.Last.lib}
\alias{my_update}
\alias{powerbasis}
\alias{invpowerbasis}
\alias{powrecbasis}
\alias{map}
\alias{invmap}
\alias{rcLoss}
\alias{L1loss}
\alias{logit}
\alias{invlogit}
\alias{cloglog}
\alias{invcloglog}
\alias{perc.ci}
\alias{norm.inter}
\alias{F.rq}
\alias{addnoise}
\alias{KhmaladzeTable}
\description{Internal Qtools objects.}
\details{These are not to be called by the user.}
\keyword{internal}