\name{Qtools-package}
\alias{Qtools-package}
\docType{package}
\title{Utilities for Quantilies}

\description{
Thi is a collection of functions for unconditional and conditional quantiles
}

\details{
\tabular{ll}{
Package: \tab Qtools\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-10-15\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}

}
\author{
Marco Geraci

Maintainer: Marco Geraci <m.geraci@ucl.ac.uk>
}

\keyword{quantiles}
