\name{residuals.midrq}
\alias{residuals.midrq}

\title{
Residuals from a midrq Objects
}

\description{
This function computes the residuals from a fitted mid-quantile regression model.
}

\usage{
\method{residuals}{midrq}(object, ...)
}

\arguments{
  \item{object}{
an \code{midrq} object.
}

  \item{\dots}{
not used.
}
}


\value{
a vector or matrix of residuals.
}


\author{
Marco Geraci
}


\seealso{
\code{\link{midrq}}
}

\keyword{residuals}
