\name{residuals.rq.counts}
\alias{residuals.rq.counts}

\title{
Residuals from an rq.counts Object
}

\description{
This function computes the residuals from a fitted linear quantile model for counts.
}

\usage{
\method{residuals}{rq.counts}(object, ...)
}

\arguments{
  \item{object}{
an \code{rq.counts} object.
}

  \item{\dots}{
not used.
}
}


\value{
a vector or matrix of residuals.
}


\author{
Marco Geraci
}


\seealso{
\code{\link{rq.counts}}
}

\keyword{residuals}
