\name{residuals.rqt}
\alias{residuals.rqt}

\title{
Residuals from an rqt Objects
}

\description{
This function computes the residuals from a fitted quantile regression transformation model.
}

\usage{
\method{residuals}{rqt}(object, ...)
}

\arguments{
  \item{object}{
an \code{rqt} object.
}

  \item{\dots}{
not used.
}
}


\value{
a vector or matrix of residuals.
}


\author{
Marco Geraci
}


\seealso{
\code{\link{tsrq}}
}

\keyword{residuals}
