\name{Qtools-internal}
\title{Internal Qtools objects}
\alias{.First.lib}
\alias{.Last.lib}

\alias{logit}
\alias{invlogit}
\alias{cloglog}
\alias{invcloglog}
\alias{powerbase}
\alias{invpowerbase}
\alias{map}
\alias{invmap}
\alias{powrecbase}
\alias{d1bc}
\alias{d1ao}
\alias{d1mcjI}
\alias{d2bc}
\alias{d2ao}
\alias{d2mcjI}

\alias{terms2expr}
\alias{boot.rrq}
\alias{boot.rqt}
\alias{se.rqt}
\alias{qlssPredRq}
\alias{qlssPredRqt}

\alias{rcLoss}
\alias{l1Loss}
\alias{nlLoss}
\alias{nl.fit.rqt}
\alias{switch_check}
\alias{addnoise}
\alias{normalize}
\alias{rcTest}
\alias{KhmaladzeTable}

\alias{C_midrqLoss}
\alias{C_midrqLoss_bc}
\alias{C_midrqLoss_ao}
\alias{C_rcTest}
\alias{C_projfun}
\alias{C_phifun}
\alias{_Qtools_C_midrqLoss}
\alias{_Qtools_C_midrqLoss_bc}
\alias{_Qtools_C_midrqLoss_ao}
\alias{_Qtools_C_rcTest}
\alias{_Qtools_C_projfun}
\alias{_Qtools_C_phifun}


\description{Internal Qtools objects.}
\details{These are not to be called by the user.}
\keyword{internal}