\name{Qtools-internal}
\title{Internal Qtools objects}
\alias{.First.lib}
\alias{.Last.lib}
\alias{marefbc}
\alias{marefao}
\alias{marefmcjI}
\alias{boot.fun}
\alias{powerbasis}
\alias{invpowerbasis}
\alias{powrecbasis}
\alias{map}
\alias{invmap}
\alias{rcLoss}
\alias{l1Loss}
\alias{logit}
\alias{invlogit}
\alias{cloglog}
\alias{invcloglog}
\alias{addnoise}
\alias{KhmaladzeTable}
\description{Internal Qtools objects.}
\details{These are not to be called by the user.}
\keyword{internal}