\name{Qtools-internal}
\title{Internal Qtools objects}
\alias{.First.lib}
\alias{.Last.lib}
\alias{d1bc}
\alias{d1ao}
\alias{d1mcjI}
\alias{d2bc}
\alias{d2ao}
\alias{d2mcjI}
\alias{sparsity}
\alias{se.rqt}
\alias{marefbc}
\alias{marefao}
\alias{marefmcjI}
\alias{boot.fun}
\alias{powerbase}
\alias{invpowerbase}
\alias{powrecbase}
\alias{map}
\alias{invmap}
\alias{rcLoss}
\alias{l1Loss}
\alias{logit}
\alias{invlogit}
\alias{cloglog}
\alias{invcloglog}
\alias{addnoise}
\alias{KhmaladzeTable}
\description{Internal Qtools objects.}
\details{These are not to be called by the user.}
\keyword{internal}