\name{Qtools-package}
\alias{Qtools-package}
\docType{package}
\title{Utilities for Quantilies}

\description{
The package Qtools provides functions for unconditional and conditional quantiles. These include methods for transformation-based quantile regression, quantile-based measures of location, scale and shape, methods for quantiles of discrete variables, quantile-based multiple imputation, and restricted quantile regression.
}

\details{
\tabular{ll}{
Package: \tab Qtools\cr
Type: \tab Package\cr
Version: \tab 1.4\cr
Date: \tab 2019-07-04\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}

}
\author{
Marco Geraci

Maintainer: Marco Geraci <geraci@mailbox.sc.edu>
}

\keyword{quantiles}
