\name{Qtools-package}
\alias{Qtools-package}
\docType{package}
\title{Utilities for quantilies}

\description{
Collection of functions for unconditional and conditional quantiles
}

\details{
\tabular{ll}{
Package: \tab Qtools\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-05-20\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}

}
\author{
Marco Geraci

Maintainer: Marco Geraci <m.geraci@ucl.ac.uk>
}

\keyword{quantiles}
